update assets if it's different (#38)

+++ reflex_chakra/components/base.py
@@ -58,13 +58,19 @@ class ChakraComponent(Component):
     def create(cls, *children, **props) -> Component:
         # copy color mode provider file to client's asset dir if it doesnt exist.
         client_asset_dir = Path.cwd() / constants.ASSETS_DIR_NAME
-        if not (
-            client_color_mode_provider := (
-                Path.cwd()
-                / constants.ASSETS_DIR_NAME
-                / constants.COLOR_MODE_PROVIDER_FILENAME
-            )
-        ).exists():
+        if (
+            not (
+                client_color_mode_provider := (
+                    Path.cwd()
+                    / constants.ASSETS_DIR_NAME
+                    / constants.COLOR_MODE_PROVIDER_FILENAME
+                )
+            ).exists()
+            or client_color_mode_provider.read_bytes()
+            != (
+                constants.ASSETS_DIR / constants.COLOR_MODE_PROVIDER_FILENAME
+            ).read_bytes()
+        ):
             client_asset_dir.mkdir(exist_ok=True)
             shutil.copy(
                 constants.ASSETS_DIR / constants.COLOR_MODE_PROVIDER_FILENAME,
